+++ stac_fastapi/extensions/stac_fastapi/extensions/core/collection_search/__init__.py
@@ -1,0 +1,5 @@
+"""Collection-Search extension module."""
+
+from .collection_search import CollectionSearchExtension, ConformanceClasses
+
+__all__ = ["CollectionSearchExtension", "ConformanceClasses"]